refactor: extract CollectionsFacade from CLIFacade
Move collection CRUD, export/import, Karabiner import, and layer CRUD
into standalone CollectionsFacade struct. CLIFacade retains only service
lifecycle, config, TCP, and pack management methods.

Closes #441
