+++ Sources/KeyPathAppKit/CLI/CLIFacade.swift
@@ -10,270 +10,11 @@ import KeyPathWizardCore
 /// Method groups live in extension files and standalone facades:
 /// - RulesFacade.swift — Custom rules CRUD (standalone)
 /// - SimulatorFacade.swift — Key simulation and validation (standalone)
-/// - CLIFacade+Collections.swift — Rule collections CRUD
+/// - CollectionsFacade.swift — Collections, export/import, layers (standalone)
 /// - CLIFacade+Service.swift — Service lifecycle, config, TCP, status, installer
 /// - CLIFacade+Packs.swift — Pack management
 public struct CLIFacade: Sendable {
     public init() {}
-
-    static func parseLayer(_ name: String) -> RuleCollectionLayer {
-        switch name.lowercased() {
-        case "base": .base
-        case "nav", "navigation": .navigation
-        default: .custom(name)
-        }
-    }
-
-    // MARK: - Export / Import
-
-    public func exportCollection(nameOrId: String) async throws -> CLIExportedCollection? {
-        let collections = await RuleCollectionStore.shared.loadCollections()
-        guard let index = try resolveCollectionIndex(nameOrId: nameOrId, in: collections) else {
-            return nil
-        }
-        return CLIExportedCollection(from: collections[index])
-    }
-
-    public func exportAllCollections() async -> [CLIExportedCollection] {
-        let collections = await RuleCollectionStore.shared.loadCollections()
-        return collections.map { CLIExportedCollection(from: $0) }
-    }
-
-    public func importCollection(_ exported: CLIExportedCollection, onConflict: CLIConflictStrategy = .fail) async throws -> CLIRuleCollection {
-        var collections = await RuleCollectionStore.shared.loadCollections()
-        let existingIndex = collections.firstIndex(where: { $0.name == exported.name })
-
-        if let existingIndex {
-            switch onConflict {
-            case .fail:
-                throw AmbiguousCollectionMatch(
-                    query: exported.name,
-                    matches: [.init(name: collections[existingIndex].name, id: collections[existingIndex].id.uuidString)],
-                    hint: "Use --on-conflict=replace to overwrite or --on-conflict=skip to no-op"
-                )
-            case .skip:
-                return CLIRuleCollection(from: collections[existingIndex])
-            case .replace, .merge:
-                collections.remove(at: existingIndex)
-            }
-        }
-
-        let collection = exported.toRuleCollection()
-        collections.append(collection)
-        try await RuleCollectionStore.shared.saveCollections(collections)
-        return CLIRuleCollection(from: collection)
-    }
-
-    // MARK: - Karabiner Import
-
-    public func importFromKarabiner(data: Data, collectionName: String?, profileIndex: Int?) throws -> CLIKarabinerImportResult {
-        let service = KarabinerConverterService()
-
-        let result: KarabinerConversionResult
-        do {
-            result = try service.convert(data: data, profileIndex: profileIndex)
-        } catch {
-            if let complexResult = try? convertComplexModsFile(data: data, service: service) {
-                result = complexResult
-            } else {
-                throw error
-            }
-        }
-
-        let exportedCollections: [CLIExportedCollection]
-        if let name = collectionName {
-            let allMappings = result.collections.flatMap(\.mappings)
-            let merged = RuleCollection(
-                name: name,
-                summary: "Imported from Karabiner profile: \(result.profileName)",
-                category: .custom,
-                mappings: allMappings
-            )
-            exportedCollections = [CLIExportedCollection(from: merged)]
-        } else {
-            exportedCollections = result.collections.map { CLIExportedCollection(from: $0) }
-        }
-
-        var warnings = result.warnings
-
-        if !result.appKeymaps.isEmpty {
-            let count = result.appKeymaps.map(\.overrides.count).reduce(0, +)
-            warnings.append("\(count) app-specific override(s) found -- use the GUI to configure app keymaps")
-        }
-
-        if !result.launcherMappings.isEmpty {
-            warnings.append("\(result.launcherMappings.count) launcher mapping(s) found -- use the GUI to configure launcher shortcuts")
-        }
-
-        let skipped = result.skippedRules.map {
-            CLISkippedRule(description: $0.description, reason: $0.reason)
-        }
-
-        return CLIKarabinerImportResult(
-            profileName: result.profileName,
-            collections: exportedCollections,
-            skippedRules: skipped,
-            warnings: warnings
-        )
-    }
-
-    public func listKarabinerProfiles(data: Data) throws -> [CLIKarabinerProfile] {
-        let service = KarabinerConverterService()
-        let profiles = try service.getProfiles(from: data)
-        return profiles.map {
-            CLIKarabinerProfile(name: $0.name, index: $0.index, isSelected: $0.isSelected)
-        }
-    }
-
-    private func convertComplexModsFile(data: Data, service: KarabinerConverterService) throws -> KarabinerConversionResult {
-        struct ComplexModsFile: Decodable {
-            let title: String?
-            let rules: [KarabinerRule]
-        }
-
-        guard let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-              json["rules"] != nil
-        else {
-            throw KarabinerImportError.invalidJSON("Not a recognized Karabiner format")
-        }
-
-        let title = json["title"] as? String ?? "Imported Rules"
-        let wrapped: [String: Any] = [
-            "profiles": [[
-                "name": title,
-                "selected": true,
-                "complex_modifications": json,
-            ]],
-        ]
-
-        let wrappedData = try JSONSerialization.data(withJSONObject: wrapped)
-        return try service.convert(data: wrappedData, profileIndex: 0)
-    }
-
-    // MARK: - Layer CRUD
-
-    public func listDefinedLayers() async -> [String] {
-        let collections = await RuleCollectionStore.shared.loadCollections()
-        var layers = Set<String>()
-        layers.insert("base")
-        for collection in collections {
-            layers.insert(collection.targetLayer.kanataName)
-        }
-        return layers.sorted()
-    }
-
-    public func createLayer(name: String) async throws -> CLIRuleCollection {
-        var collections = await RuleCollectionStore.shared.loadCollections()
-        let layer: RuleCollectionLayer = switch name.lowercased() {
-        case "base": .base
-        case "nav", "navigation": .navigation
-        default: .custom(name)
-        }
-        let collection = RuleCollection(
-            name: "\(name) Layer",
-            summary: "Rules for the \(name) layer",
-            category: .layers,
-            mappings: [],
-            targetLayer: layer
-        )
-        collections.append(collection)
-        try await RuleCollectionStore.shared.saveCollections(collections)
-        return CLIRuleCollection(from: collection)
-    }
-
-    public func deleteLayer(name: String) async throws -> Int {
-        var collections = await RuleCollectionStore.shared.loadCollections()
-        let targetName = Self.parseLayer(name).kanataName
-        let before = collections.count
-        collections.removeAll { $0.targetLayer.kanataName == targetName }
-        let removed = before - collections.count
-        if removed > 0 {
-            try await RuleCollectionStore.shared.saveCollections(collections)
-        }
-        return removed
-    }
-
-    public func renameLayer(oldName: String, newName: String) async throws -> Int {
-        var collections = await RuleCollectionStore.shared.loadCollections()
-        let oldLayerName = Self.parseLayer(oldName).kanataName
-        let newLayer = Self.parseLayer(newName)
-        var updated = 0
-        for i in collections.indices {
-            if collections[i].targetLayer.kanataName == oldLayerName {
-                collections[i].targetLayer = newLayer
-                updated += 1
-            }
-        }
-        if updated > 0 {
-            try await RuleCollectionStore.shared.saveCollections(collections)
-        }
-        return updated
-    }
-
-    // MARK: - Helpers
-
-    func resolveCollectionIndex(nameOrId: String, in collections: [RuleCollection]) throws -> Int? {
-        if let index = collections.firstIndex(where: { $0.id.uuidString == nameOrId }) {
-            return index
-        }
-
-        let exactMatches = collections.enumerated().filter {
-            $0.element.name.caseInsensitiveCompare(nameOrId) == .orderedSame
-        }
-        if exactMatches.count == 1 {
-            return exactMatches[0].offset
-        }
-        if exactMatches.count > 1 {
-            throw AmbiguousCollectionMatch(
-                query: nameOrId,
-                matches: exactMatches.map { .init(name: $0.element.name, id: $0.element.id.uuidString) },
-                hint: "Multiple collections share this name. Use the ID to disambiguate."
-            )
-        }
-
-        let substringMatches = collections.enumerated().filter {
-            $0.element.name.localizedCaseInsensitiveContains(nameOrId)
-        }
-        if substringMatches.count == 1 {
-            return substringMatches[0].offset
-        }
-        if substringMatches.count > 1 {
-            throw AmbiguousCollectionMatch(
-                query: nameOrId,
-                matches: substringMatches.map { .init(name: $0.element.name, id: $0.element.id.uuidString) }
-            )
-        }
-
-        return nil
-    }
-}
-
-// MARK: - Ambiguous Match Errors
-
-public struct AmbiguousCollectionMatch: Error, CustomStringConvertible {
-    public struct Match: Sendable {
-        public let name: String
-        public let id: String
-    }
-
-    public let query: String
-    public let matches: [Match]
-    public let hint: String
-
-    public init(query: String, matches: [Match], hint: String = "Use the full name or ID to disambiguate.") {
-        self.query = query
-        self.matches = matches
-        self.hint = hint
-    }
-
-    public var description: String {
-        var lines = ["Found \(matches.count) collections matching \"\(query)\":"]
-        for match in matches {
-            lines.append("  - \(match.name) (id: \(match.id))")
-        }
-        lines.append(hint)
-        return lines.joined(separator: "\n")
-    }
 }
 
 // MARK: - Version
@@ -295,23 +36,7 @@ public enum CLIVersion {
     }()
 }
 
-// MARK: - Public CLI Types
-
-public struct CLIRuleCollection: Codable, Sendable {
-    public let id: String
-    public let name: String
-    public let isEnabled: Bool
-    public let mappingCount: Int
-    public let summary: String
-
-    public init(from collection: RuleCollection) {
-        id = collection.id.uuidString
-        name = collection.name
-        isEnabled = collection.isEnabled
-        mappingCount = collection.mappings.count
-        summary = collection.summary
-    }
-}
+// MARK: - Service & Config Types
 
 public struct CLIApplyResult: Codable, Sendable {
     public let collectionsCount: Int
@@ -407,81 +132,6 @@ public struct CLIInspectResult: Codable, Sendable {
     public let planStatus: String
     public let blockedBy: String?
     public let plannedRecipes: [String]
-}
-
-// MARK: - Karabiner Import Types
-
-public struct CLIKarabinerImportResult: Codable, Sendable {
-    public let profileName: String
-    public let collections: [CLIExportedCollection]
-    public let skippedRules: [CLISkippedRule]
-    public let warnings: [String]
-}
-
-public struct CLISkippedRule: Codable, Sendable {
-    public let description: String
-    public let reason: String
-}
-
-public struct CLIKarabinerProfile: Codable, Sendable {
-    public let name: String
-    public let index: Int
-    public let isSelected: Bool
-}
-
-// MARK: - Export/Import Types
-
-public struct CLIExportedCollection: Codable, Sendable {
-    public let name: String
-    public let summary: String
-    public let category: String
-    public let isEnabled: Bool
-    public let targetLayer: String
-    public let mappings: [CLIExportedMapping]
-
-    public init(from collection: RuleCollection) {
-        name = collection.name
-        summary = collection.summary
-        category = collection.category.rawValue
-        isEnabled = collection.isEnabled
-        targetLayer = collection.targetLayer.kanataName
-        mappings = collection.mappings.map { CLIExportedMapping(from: $0) }
-    }
-
-    public func toRuleCollection() -> RuleCollection {
-        let cat = RuleCollectionCategory(rawValue: category) ?? .custom
-        let layer: RuleCollectionLayer = switch targetLayer {
-        case "base": .base
-        case "nav": .navigation
-        default: .custom(targetLayer)
-        }
-        return RuleCollection(
-            name: name,
-            summary: summary,
-            category: cat,
-            mappings: mappings.map { $0.toKeyMapping() },
-            isEnabled: isEnabled,
-            targetLayer: layer
-        )
-    }
-}
-
-public struct CLIExportedMapping: Codable, Sendable {
-    public let input: String
-    public let action: KeyAction
-    public let shiftedOutput: String?
-    public let behavior: MappingBehavior?
-
-    public init(from mapping: KeyMapping) {
-        input = mapping.input
-        action = mapping.action
-        shiftedOutput = mapping.shiftedOutput
-        behavior = mapping.behavior
-    }
-
-    public func toKeyMapping() -> KeyMapping {
-        KeyMapping(input: input, action: action, shiftedOutput: shiftedOutput, behavior: behavior)
-    }
 }
 
 // MARK: - Pack CLI Types
