+++ Sources/KeyPathCLI/Commands/Collection/CollectionDisableCommand.swift
@@ -18,7 +18,7 @@ struct CollectionDisable: AsyncParsableCommand {
 
     mutating func run() async throws {
         let ctx = globals.outputContext
-        let facade = await MainActor.run { CLIFacade() }
+        let facade = CollectionsFacade()
 
         do {
             guard let name = try await facade.disableCollection(nameOrId: nameOrId) else {
