+++ Sources/KeyPathCLI/Commands/Collection/CollectionRenameCommand.swift
@@ -18,7 +18,7 @@ struct CollectionRename: AsyncParsableCommand {
 
     mutating func run() async throws {
         let ctx = globals.outputContext
-        let facade = await MainActor.run { CLIFacade() }
+        let facade = CollectionsFacade()
 
         do {
             guard let oldName = try await facade.renameCollection(nameOrId: nameOrId, newName: newName) else {
