+++ Sources/KeyPathCLI/Commands/Collection/CollectionShowCommand.swift
@@ -15,7 +15,7 @@ struct CollectionShow: AsyncParsableCommand {
 
     mutating func run() async throws {
         let ctx = globals.outputContext
-        let facade = await MainActor.run { CLIFacade() }
+        let facade = CollectionsFacade()
 
         let collection: CLIRuleCollection
         do {
