+++ Sources/KeyPathCLI/Commands/Export/ExportAllCommand.swift
@@ -15,7 +15,7 @@ struct ExportAll: AsyncParsableCommand {
 
     mutating func run() async throws {
         let ctx = globals.outputContext
-        let facade = await MainActor.run { CLIFacade() }
+        let facade = CollectionsFacade()
 
         let exported = await facade.exportAllCollections()
