+++ Sources/KeyPathCLI/Commands/Export/ExportCollectionCommand.swift
@@ -18,7 +18,7 @@ struct ExportCollection: AsyncParsableCommand {
 
     mutating func run() async throws {
         let ctx = globals.outputContext
-        let facade = await MainActor.run { CLIFacade() }
+        let facade = CollectionsFacade()
 
         do {
             guard let exported = try await facade.exportCollection(nameOrId: nameOrId) else {
