+++ Sources/KeyPathCLI/Commands/Import/ImportKarabinerCommand.swift
@@ -21,7 +21,7 @@ struct ImportKarabiner: AsyncParsableCommand {
 
     mutating func run() async throws {
         let ctx = globals.outputContext
-        let facade = await MainActor.run { CLIFacade() }
+        let facade = CollectionsFacade()
 
         let url = URL(fileURLWithPath: (path as NSString).expandingTildeInPath)
         let data: Data
