+++ Sources/KeyPathCLI/Commands/Layer/LayerCreateCommand.swift
@@ -15,7 +15,7 @@ struct LayerCreate: AsyncParsableCommand {
 
     mutating func run() async throws {
         let ctx = globals.outputContext
-        let facade = await MainActor.run { CLIFacade() }
+        let facade = CollectionsFacade()
 
         let collection = try await facade.createLayer(name: name)
         CLIOutput.write(collection, context: ctx) {
