+++ Tests/KeyPathTests/CLI/CLIKarabinerImportTests.swift
@@ -3,7 +3,7 @@
 
 @MainActor
 final class CLIKarabinerImportTests: XCTestCase {
-    private let facade = CLIFacade()
+    private let facade = CollectionsFacade()
     private var originalCollections: [RuleCollection] = []
 
     override func setUp() async throws {
